examples: fix examples/sage
