+++ config/.prettierignore
@@ -1,5 +1,4 @@
 ../**/bin
-../**/*.json
 ../**/node_modules
 ../**/package.json
 ../**/tsconfig.json
