+++ examples/sage/resources/images/image.jpeg
