+++ examples/sage/resources/images/nested/image.jpeg
