+++ examples/sage/resources/scripts/app.js
@@ -1,0 +1,10 @@
+import {main} from '@scripts/components/main'
+
+const init = () =>
+  window.requestAnimationFrame(function ready() {
+    return document.body ? main() : window.requestAnimationFrame(ready)
+  })
+
+init()
+
+module?.hot?.accept('./components/main.js', init)
