+++ examples/sage/resources/scripts/components/main.js
@@ -7,5 +7,5 @@ export const main = () => {
    * Remove no-js class from body element
    */
   document.body.classList.contains('no-js') &&
-    document.body.classList.remove('no-js');
-};
+    document.body.classList.remove('no-js')
+}
