+++ examples/sage/resources/scripts/customizer.js
@@ -1,0 +1,5 @@
+wp.customize('blogname', value => {
+  value.bind(to => {
+    $('.brand').text(to)
+  })
+})
