+++ examples/sage/resources/styles/common/global.css
