+++ examples/sage/resources/styles/editor.css
