--- examples/sage/src/scripts/customizer.js
@@ -1,5 +1,0 @@
-wp.customize('blogname', (value) => {
-  value.bind((to) => {
-    $('.brand').text(to);
-  });
-});
